Refactoring (#99)
* refactor: eslint and typechecking

* refactor: forge fmt on pool contracts

* refactor: forge fmt files in base contract dir

* refactor: bonding curves

- forge fmt 2nd pass on previously refactored contracts

* refactor: forge fmt filter directory

* refactor: forge mt rng dir

* refactor: forge fmt validators

* feat: package.json yarn format script

* feat: yarn lint script lints both ts and sol files

* refactor: pools. routers, vaults directories, tree

- contracts directory should be partitioned to self document contracts

* refactor: resolve foundry test imports after reorg

* fix: update yarn format with new paths

- {path1,path2} syntax not native for yarn. pass as bash input

* refactor: test contracts go into contracts/test

* feat: yarn deploy and verify scripts

- test contracts only compiled when running yarn test:hh
- remove tree from deploy/verify script
- new deploy to mumbai to show reduced set of contracts verified
- minor lint in fixtures.ts

* refactor: final lints due to rebase conflicts

* doc: README update to reflect yarn scripts

* fix: revert incorrect conflict resolution

- ivalidator and monotonic validator royaltynumerator were missing
- decouple format from lint

* fix: import paths for foundry tests

* fix: update github workflow

* fix: remove forge test from workflow

- refactor variable name in hardhat config
- eslintrc disable capitalized-comments rule
- reverted hardhat comment lint

+++ contracts/filter/TokenIDFilter.sol
@@ -12,18 +12,12 @@ contract TokenIDFilter is ITokenIDFilter {
 
     uint32[999] _padding;
 
-    function _setRootAndEmitAcceptedIDs(
-        address collection,
-        bytes32 root,
-        bytes calldata data
-    ) internal {
+    function _setRootAndEmitAcceptedIDs(address collection, bytes32 root, bytes calldata data) internal {
         tokenIDFilterRoot = root;
         emit AcceptsTokenIDs(collection, tokenIDFilterRoot, data);
     }
 
-    function _acceptsTokenID(uint256 tokenID, bytes32[] calldata proof)
-    internal view returns (bool)
-    {
+    function _acceptsTokenID(uint256 tokenID, bytes32[] calldata proof) internal view returns (bool) {
         if (tokenIDFilterRoot == 0) {
             return true;
         }
@@ -39,19 +33,23 @@ contract TokenIDFilter is ITokenIDFilter {
     }
 
     function _acceptsTokenIDs(uint256[] calldata tokenIDs, bytes32[] calldata proof, bool[] calldata proofFlags)
-    internal view returns (bool)
+        internal
+        view
+        returns (bool)
     {
         if (tokenIDFilterRoot == 0) {
             return true;
         }
 
-        uint length = tokenIDs.length;
+        uint256 length = tokenIDs.length;
         bytes32[] memory leaves = new bytes32[](length);
 
-        for (uint i; i < length;) {
+        for (uint256 i; i < length;) {
             // double hash to prevent second preimage attack
             leaves[i] = keccak256(abi.encodePacked(keccak256(abi.encodePacked((tokenIDs[i])))));
-            unchecked { ++i; }
+            unchecked {
+                ++i;
+            }
         }
 
         return MerkleProof.multiProofVerify(proof, proofFlags, tokenIDFilterRoot, leaves);
